Pull ArtifactsCache out of artifacts.py
