+++ wandb/sdk/interface/artifacts/__init__.py
@@ -1,14 +1,20 @@
-from .artifacts import (
+from wandb.sdk.interface.artifacts.artifacts import (
     Artifact,
     ArtifactManifest,
     ArtifactManifestEntry,
     ArtifactNotLoggedError,
-    ArtifactsCache,
-    get_artifacts_cache,
     get_new_staging_file,
     get_staging_dir,
 )
-from .artifacts_storage import StorageHandler, StorageLayout, StoragePolicy
+from wandb.sdk.interface.artifacts.artifacts_cache import (
+    ArtifactsCache,
+    get_artifacts_cache,
+)
+from wandb.sdk.interface.artifacts.artifacts_storage import (
+    StorageHandler,
+    StorageLayout,
+    StoragePolicy,
+)
 
 __all__ = [
     "Artifact",
